+++ src/main/java/nordmods/iobvariantloader/mixin/common/egg/DragonEggItemMixin.java
@@ -2,12 +2,7 @@ package nordmods.iobvariantloader.mixin.common.egg;
 
 import com.GACMD.isleofberk.entity.eggs.entity.base.ADragonEggBase;
 import com.GACMD.isleofberk.items.DragonEggItem;
-import net.minecraft.ChatFormatting;
-import net.minecraft.locale.Language;
 import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.Style;
-import net.minecraft.network.chat.TextComponent;
-import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.EntityType;
@@ -19,15 +14,15 @@ import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.ServerLevelAccessor;
 import nordmods.iobvariantloader.IoBVariantLoader;
-import nordmods.iobvariantloader.util.ducks.DragonSpeciesHelper;
-import nordmods.iobvariantloader.util.ducks.VariantNameHelper;
+import nordmods.iobvariantloader.util.ResourceUtil;
 import nordmods.iobvariantloader.util.dragon_variant_spawner.DragonVariantSpawner;
 import nordmods.iobvariantloader.util.dragon_variant_spawner.DragonVariantSpawnerUtil;
+import nordmods.iobvariantloader.util.ducks.DragonSpeciesHelper;
+import nordmods.iobvariantloader.util.ducks.VariantNameHelper;
 import nordmods.iobvariantloader.util.model_redirect.ModelRedirectUtil;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -60,32 +55,7 @@ public abstract class DragonEggItemMixin extends Item implements DragonSpeciesHe
     @Inject(method = "appendHoverText(Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/level/Level;Ljava/util/List;Lnet/minecraft/world/item/TooltipFlag;)V", at = @At("TAIL"))
     private void addVariantTooltip(ItemStack pStack, Level pLevel, List<Component> pTooltipComponents, TooltipFlag pIsAdvanced, CallbackInfo ci) {
         String variant = pStack.hasTag() ? pStack.getTag().getString("VariantName") : "";
-        if (!variant.isEmpty()) {
-            String key = "tooltip.iobvariantloader." + getSpecies(true) + "." + variant;
-            if (Language.getInstance().has(key)) {
-                pTooltipComponents.add(new TranslatableComponent("tooltip.iobvariantloader.variant", new TranslatableComponent(key).withStyle(ChatFormatting.GOLD)));
-            } else {
-                variant = parseName(variant);
-                pTooltipComponents.add(new TranslatableComponent("tooltip.iobvariantloader.variant", new TextComponent(variant).withStyle(ChatFormatting.GOLD)));
-            }
-        } else pTooltipComponents.add(new TranslatableComponent("tooltip.iobvariantloader.variant",
-                new TextComponent("unknown").setStyle(Style.EMPTY.withColor(ChatFormatting.GOLD).withObfuscated(true))));
-    }
-
-    @Unique
-    private String parseName(String name) {
-        while (name.contains("_")) {
-            int index = name.indexOf("_");
-            if (index + 1 < name.length()) {
-                String toReplace = String.valueOf(name.charAt(index + 1));
-                name = name.replaceFirst("_" + toReplace, " " + toReplace.toUpperCase());
-                continue;
-            }
-            name = name.replace("_", " ");
-        }
-        name = name.replace(" N ", "'n'");
-        String firstLetter = String.valueOf(name.charAt(0));
-        return name.replaceFirst(firstLetter, firstLetter.toUpperCase());
+        pTooltipComponents.add(ResourceUtil.getVariantNameTooltip(variant, this));
     }
 
     @Override